refactor(goals): extract lifecycleProjection.ts pure functions (step 10)
Move buildLifecycle, buildPlannedProjection, and buildProjectedLifecycle
into src/pages/goal/utils/lifecycleProjection.ts. GoalDiveDeep now imports
from the shared module. ~180 lines removed from component.
